Feature/pedido calcado

+++ backend/app/Http/Controllers/UsuarioController.php
@@ -9,37 +9,33 @@
   
  class UsuarioController extends Controller
  {
-    /**
-     * Lista de todos os usuários.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index(){
-        $usuarios = Usuario::all();
-        return $usuarios;
-    }
-  
-    /**
-     * Armazena um novo usuário.
-     *
-     */
-    public function store(Request $request){
-        $usu = $request->all();
-        $usuario = Usuario::create($usu);
-        return $usuario;
+  public function index(){
+    return Usuario::all();
+  }
+
+  public function store(Request $request){
+    $usuario = Usuario::create($request->all());
+
+    return response()->json($usuario);
+  }
+
+  public function show(Usuario $usuario) {
+    return response()->json($usuario);
+  }
+
+  public function update(Request $request, Usuario $usuario) {
+    if ($usuario == auth()->user()) {
+      $usuario->update($request->all());
     }
 
-    public function show(Usuario $usuario) {
-        return response()->json($usuario);
-    }
-  
-    public function update(Request $request, Usuario $usuario) {
-        $usuario->update($request->all());
-        return response()->json($usuario, 200);
-    }
-  
-    public function delete(Usuario $usuario) {
-        $usuario->delete();
+    return response()->json($usuario, 200);
+  }
+
+  public function delete(Usuario $usuario) {
+    if ($usuario == auth()->user()) {
+      $usuario->delete();
     }
 
+    return response('', 204);
+  }
  }
